requirements.txt
bert_extractive_summarizer==0.10.1
nltk==3.8.1
PyYAML==6.0.2
Requests==2.32.3
spacy==3.5.0
sumy==0.11.0
torch==2.3.1
transformers==4.42.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl